move tests from queue.rs to state.rs
The queue.rs file is going to be deleted as we're going to remove the
Queue implementation. The tests are already using the QueueState
structure, so we can just move them directly to the corresponding file.
