+++ crates/virtio-queue/src/state.rs
@@ -485,3 +485,471 @@ impl QueueStateOwnedT for QueueState {
         self.next_avail -= Wrapping(1);
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::defs::{
+        DEFAULT_AVAIL_RING_ADDR, DEFAULT_DESC_TABLE_ADDR, DEFAULT_USED_RING_ADDR,
+        VIRTQ_DESC_F_NEXT, VIRTQ_DESC_F_WRITE, VIRTQ_USED_F_NO_NOTIFY,
+    };
+    use crate::mock::MockSplitQueue;
+    use crate::Descriptor;
+
+    use vm_memory::{Address, Bytes, GuestAddress, GuestMemoryMmap};
+
+    #[test]
+    fn test_queue_is_valid() {
+        let m = &GuestMemoryMmap::<()>::from_ranges(&[(GuestAddress(0), 0x10000)]).unwrap();
+        let vq = MockSplitQueue::new(m, 16);
+        let mut q: QueueState = vq.create_queue();
+
+        // q is currently valid
+        assert!(q.is_valid(m));
+
+        // shouldn't be valid when not marked as ready
+        q.set_ready(false);
+        assert!(!q.ready());
+        assert!(!q.is_valid(m));
+        q.set_ready(true);
+
+        // shouldn't be allowed to set a size > max_size
+        q.set_size(q.max_size() << 1);
+        assert_eq!(q.size, q.max_size());
+
+        // or set the size to 0
+        q.set_size(0);
+        assert_eq!(q.size, q.max_size());
+
+        // or set a size which is not a power of 2
+        q.set_size(11);
+        assert_eq!(q.size, q.max_size());
+
+        // but should be allowed to set a size if 0 < size <= max_size and size is a power of two
+        q.set_size(4);
+        assert_eq!(q.size, 4);
+        q.size = q.max_size();
+
+        // shouldn't be allowed to set an address that breaks the alignment constraint
+        q.set_desc_table_address(Some(0xf), None);
+        assert_eq!(q.desc_table.0, vq.desc_table_addr().0);
+        // should be allowed to set an aligned out of bounds address
+        q.set_desc_table_address(Some(0xffff_fff0), None);
+        assert_eq!(q.desc_table.0, 0xffff_fff0);
+        // but shouldn't be valid
+        assert!(!q.is_valid(m));
+        // but should be allowed to set a valid description table address
+        q.set_desc_table_address(Some(0x10), None);
+        assert_eq!(q.desc_table.0, 0x10);
+        assert!(q.is_valid(m));
+        let addr = vq.desc_table_addr().0;
+        q.set_desc_table_address(Some(addr as u32), Some((addr >> 32) as u32));
+
+        // shouldn't be allowed to set an address that breaks the alignment constraint
+        q.set_avail_ring_address(Some(0x1), None);
+        assert_eq!(q.avail_ring.0, vq.avail_addr().0);
+        // should be allowed to set an aligned out of bounds address
+        q.set_avail_ring_address(Some(0xffff_fffe), None);
+        assert_eq!(q.avail_ring.0, 0xffff_fffe);
+        // but shouldn't be valid
+        assert!(!q.is_valid(m));
+        // but should be allowed to set a valid available ring address
+        q.set_avail_ring_address(Some(0x2), None);
+        assert_eq!(q.avail_ring.0, 0x2);
+        assert!(q.is_valid(m));
+        let addr = vq.avail_addr().0;
+        q.set_avail_ring_address(Some(addr as u32), Some((addr >> 32) as u32));
+
+        // shouldn't be allowed to set an address that breaks the alignment constraint
+        q.set_used_ring_address(Some(0x3), None);
+        assert_eq!(q.used_ring.0, vq.used_addr().0);
+        // should be allowed to set an aligned out of bounds address
+        q.set_used_ring_address(Some(0xffff_fffc), None);
+        assert_eq!(q.used_ring.0, 0xffff_fffc);
+        // but shouldn't be valid
+        assert!(!q.is_valid(m));
+        // but should be allowed to set a valid used ring address
+        q.set_used_ring_address(Some(0x4), None);
+        assert_eq!(q.used_ring.0, 0x4);
+        let addr = vq.used_addr().0;
+        q.set_used_ring_address(Some(addr as u32), Some((addr >> 32) as u32));
+        assert!(q.is_valid(m));
+    }
+
+    #[test]
+    fn test_add_used() {
+        let mem = &GuestMemoryMmap::<()>::from_ranges(&[(GuestAddress(0), 0x10000)]).unwrap();
+        let vq = MockSplitQueue::new(mem, 16);
+        let mut q: QueueState = vq.create_queue();
+
+        assert_eq!(q.used_idx(mem, Ordering::Acquire).unwrap(), Wrapping(0));
+        assert_eq!(u16::from_le(vq.used().idx().load()), 0);
+
+        // index too large
+        assert!(q.add_used(mem, 16, 0x1000).is_err());
+        assert_eq!(u16::from_le(vq.used().idx().load()), 0);
+
+        // should be ok
+        q.add_used(mem, 1, 0x1000).unwrap();
+        assert_eq!(q.next_used, Wrapping(1));
+        assert_eq!(q.used_idx(mem, Ordering::Acquire).unwrap(), Wrapping(1));
+        assert_eq!(u16::from_le(vq.used().idx().load()), 1);
+
+        let x = vq.used().ring().ref_at(0).unwrap().load();
+        assert_eq!(x.id(), 1);
+        assert_eq!(x.len(), 0x1000);
+    }
+
+    #[test]
+    fn test_reset_queue() {
+        let m = &GuestMemoryMmap::<()>::from_ranges(&[(GuestAddress(0), 0x10000)]).unwrap();
+        let vq = MockSplitQueue::new(m, 16);
+        let mut q: QueueState = vq.create_queue();
+
+        q.set_size(8);
+        // The address set by `MockSplitQueue` for the descriptor table is DEFAULT_DESC_TABLE_ADDR,
+        // so let's change it for testing the reset.
+        q.set_desc_table_address(Some(0x5000), None);
+        // Same for `event_idx_enabled`, `next_avail` `next_used` and `signalled_used`.
+        q.set_event_idx(true);
+        q.set_next_avail(2);
+        q.set_next_used(4);
+        q.num_added = Wrapping(15);
+        assert_eq!(q.size, 8);
+        // `create_queue` also marks the queue as ready.
+        assert!(q.ready);
+        assert_ne!(q.desc_table, GuestAddress(DEFAULT_DESC_TABLE_ADDR));
+        assert_ne!(q.avail_ring, GuestAddress(DEFAULT_AVAIL_RING_ADDR));
+        assert_ne!(q.used_ring, GuestAddress(DEFAULT_USED_RING_ADDR));
+        assert_ne!(q.next_avail, Wrapping(0));
+        assert_ne!(q.next_used, Wrapping(0));
+        assert_ne!(q.num_added, Wrapping(0));
+        assert!(q.event_idx_enabled);
+
+        q.reset();
+        assert_eq!(q.size, 16);
+        assert!(!q.ready);
+        assert_eq!(q.desc_table, GuestAddress(DEFAULT_DESC_TABLE_ADDR));
+        assert_eq!(q.avail_ring, GuestAddress(DEFAULT_AVAIL_RING_ADDR));
+        assert_eq!(q.used_ring, GuestAddress(DEFAULT_USED_RING_ADDR));
+        assert_eq!(q.next_avail, Wrapping(0));
+        assert_eq!(q.next_used, Wrapping(0));
+        assert_eq!(q.num_added, Wrapping(0));
+        assert!(!q.event_idx_enabled);
+    }
+
+    #[test]
+    fn test_needs_notification() {
+        let mem = &GuestMemoryMmap::<()>::from_ranges(&[(GuestAddress(0), 0x10000)]).unwrap();
+        let qsize = 16;
+        let vq = MockSplitQueue::new(mem, qsize);
+        let mut q: QueueState = vq.create_queue();
+        let avail_addr = vq.avail_addr();
+
+        // It should always return true when EVENT_IDX isn't enabled.
+        for i in 0..qsize {
+            q.next_used = Wrapping(i);
+            assert!(q.needs_notification(mem).unwrap());
+        }
+
+        mem.write_obj::<u16>(
+            u16::to_le(4),
+            avail_addr.unchecked_add(4 + qsize as u64 * 2),
+        )
+        .unwrap();
+        q.set_event_idx(true);
+
+        // Incrementing up to this value causes an `u16` to wrap back to 0.
+        let wrap = u32::from(u16::MAX) + 1;
+
+        for i in 0..wrap + 12 {
+            q.next_used = Wrapping(i as u16);
+            // Let's test wrapping around the maximum index value as well.
+            // `num_added` needs to be at least `1` to represent the fact that new descriptor
+            // chains have be added to the used ring since the last time `needs_notification`
+            // returned.
+            q.num_added = Wrapping(1);
+            let expected = i == 5 || i == (5 + wrap);
+            assert_eq!((q.needs_notification(mem).unwrap(), i), (expected, i));
+        }
+
+        mem.write_obj::<u16>(
+            u16::to_le(8),
+            avail_addr.unchecked_add(4 + qsize as u64 * 2),
+        )
+        .unwrap();
+
+        // Returns `false` because the current `used_event` value is behind both `next_used` and
+        // the value of `next_used` at the time when `needs_notification` last returned (which is
+        // computed based on `num_added` as described in the comments for `needs_notification`.
+        assert!(!q.needs_notification(mem).unwrap());
+
+        mem.write_obj::<u16>(
+            u16::to_le(15),
+            avail_addr.unchecked_add(4 + qsize as u64 * 2),
+        )
+        .unwrap();
+
+        q.num_added = Wrapping(1);
+        assert!(!q.needs_notification(mem).unwrap());
+
+        q.next_used = Wrapping(15);
+        q.num_added = Wrapping(1);
+        assert!(!q.needs_notification(mem).unwrap());
+
+        q.next_used = Wrapping(16);
+        q.num_added = Wrapping(1);
+        assert!(q.needs_notification(mem).unwrap());
+
+        // Calling `needs_notification` again immediately returns `false`.
+        assert!(!q.needs_notification(mem).unwrap());
+
+        mem.write_obj::<u16>(
+            u16::to_le(u16::MAX - 3),
+            avail_addr.unchecked_add(4 + qsize as u64 * 2),
+        )
+        .unwrap();
+        q.next_used = Wrapping(u16::MAX - 2);
+        q.num_added = Wrapping(1);
+        // Returns `true` because, when looking at circular sequence of indices of the used ring,
+        // the value we wrote in the `used_event` appears between the "old" value of `next_used`
+        // (i.e. `next_used` - `num_added`) and the current `next_used`, thus suggesting that we
+        // need to notify the driver.
+        assert!(q.needs_notification(mem).unwrap());
+    }
+
+    #[test]
+    fn test_enable_disable_notification() {
+        let mem = &GuestMemoryMmap::<()>::from_ranges(&[(GuestAddress(0), 0x10000)]).unwrap();
+        let vq = MockSplitQueue::new(mem, 16);
+
+        let mut q: QueueState = vq.create_queue();
+        let used_addr = vq.used_addr();
+
+        assert!(!q.event_idx_enabled);
+
+        q.enable_notification(mem).unwrap();
+        let v = mem.read_obj::<u16>(used_addr).map(u16::from_le).unwrap();
+        assert_eq!(v, 0);
+
+        q.disable_notification(mem).unwrap();
+        let v = mem.read_obj::<u16>(used_addr).map(u16::from_le).unwrap();
+        assert_eq!(v, VIRTQ_USED_F_NO_NOTIFY);
+
+        q.enable_notification(mem).unwrap();
+        let v = mem.read_obj::<u16>(used_addr).map(u16::from_le).unwrap();
+        assert_eq!(v, 0);
+
+        q.set_event_idx(true);
+        let avail_addr = vq.avail_addr();
+        mem.write_obj::<u16>(u16::to_le(2), avail_addr.unchecked_add(2))
+            .unwrap();
+
+        assert!(q.enable_notification(mem).unwrap());
+        q.next_avail = Wrapping(2);
+        assert!(!q.enable_notification(mem).unwrap());
+
+        mem.write_obj::<u16>(u16::to_le(8), avail_addr.unchecked_add(2))
+            .unwrap();
+
+        assert!(q.enable_notification(mem).unwrap());
+        q.next_avail = Wrapping(8);
+        assert!(!q.enable_notification(mem).unwrap());
+    }
+
+    #[test]
+    fn test_consume_chains_with_notif() {
+        let mem = &GuestMemoryMmap::<()>::from_ranges(&[(GuestAddress(0), 0x10000)]).unwrap();
+        let vq = MockSplitQueue::new(mem, 16);
+
+        let mut q: QueueState = vq.create_queue();
+
+        // q is currently valid.
+        assert!(q.is_valid(mem));
+
+        // The chains are (0, 1), (2, 3, 4), (5, 6), (7, 8), (9, 10, 11, 12).
+        for i in 0..13 {
+            let flags = match i {
+                1 | 4 | 6 | 8 | 12 => 0,
+                _ => VIRTQ_DESC_F_NEXT,
+            };
+
+            let desc = Descriptor::new((0x1000 * (i + 1)) as u64, 0x1000, flags, i + 1);
+            vq.desc_table().store(i, desc).unwrap();
+        }
+
+        vq.avail().ring().ref_at(0).unwrap().store(u16::to_le(0));
+        vq.avail().ring().ref_at(1).unwrap().store(u16::to_le(2));
+        vq.avail().ring().ref_at(2).unwrap().store(u16::to_le(5));
+        vq.avail().ring().ref_at(3).unwrap().store(u16::to_le(7));
+        vq.avail().ring().ref_at(4).unwrap().store(u16::to_le(9));
+        // Let the device know it can consume chains with the index < 2.
+        vq.avail().idx().store(u16::to_le(2));
+        // No descriptor chains are consumed at this point.
+        assert_eq!(q.next_avail(), 0);
+
+        let mut i = 0;
+
+        loop {
+            i += 1;
+            q.disable_notification(mem).unwrap();
+
+            while let Some(chain) = q.iter(mem).unwrap().next() {
+                // Process the descriptor chain, and then add entries to the
+                // used ring.
+                let head_index = chain.head_index();
+                let mut desc_len = 0;
+                chain.for_each(|d| {
+                    if d.flags() & VIRTQ_DESC_F_WRITE == VIRTQ_DESC_F_WRITE {
+                        desc_len += d.len();
+                    }
+                });
+                q.add_used(mem, head_index, desc_len).unwrap();
+            }
+            if !q.enable_notification(mem).unwrap() {
+                break;
+            }
+        }
+        // The chains should be consumed in a single loop iteration because there's nothing updating
+        // the `idx` field of the available ring in the meantime.
+        assert_eq!(i, 1);
+        // The next chain that can be consumed should have index 2.
+        assert_eq!(q.next_avail(), 2);
+        assert_eq!(q.next_used(), 2);
+        // Let the device know it can consume one more chain.
+        vq.avail().idx().store(u16::to_le(3));
+        i = 0;
+
+        loop {
+            i += 1;
+            q.disable_notification(mem).unwrap();
+
+            while let Some(chain) = q.iter(mem).unwrap().next() {
+                // Process the descriptor chain, and then add entries to the
+                // used ring.
+                let head_index = chain.head_index();
+                let mut desc_len = 0;
+                chain.for_each(|d| {
+                    if d.flags() & VIRTQ_DESC_F_WRITE == VIRTQ_DESC_F_WRITE {
+                        desc_len += d.len();
+                    }
+                });
+                q.add_used(mem, head_index, desc_len).unwrap();
+            }
+
+            // For the simplicity of the test we are updating here the `idx` value of the available
+            // ring. Ideally this should be done on a separate thread.
+            // Because of this update, the loop should be iterated again to consume the new
+            // available descriptor chains.
+            vq.avail().idx().store(u16::to_le(4));
+            if !q.enable_notification(mem).unwrap() {
+                break;
+            }
+        }
+        assert_eq!(i, 2);
+        // The next chain that can be consumed should have index 4.
+        assert_eq!(q.next_avail(), 4);
+        assert_eq!(q.next_used(), 4);
+
+        // Set an `idx` that is bigger than the number of entries added in the ring.
+        // This is an allowed scenario, but the indexes of the chain will have unexpected values.
+        vq.avail().idx().store(u16::to_le(7));
+        loop {
+            q.disable_notification(mem).unwrap();
+
+            while let Some(chain) = q.iter(mem).unwrap().next() {
+                // Process the descriptor chain, and then add entries to the
+                // used ring.
+                let head_index = chain.head_index();
+                let mut desc_len = 0;
+                chain.for_each(|d| {
+                    if d.flags() & VIRTQ_DESC_F_WRITE == VIRTQ_DESC_F_WRITE {
+                        desc_len += d.len();
+                    }
+                });
+                q.add_used(mem, head_index, desc_len).unwrap();
+            }
+            if !q.enable_notification(mem).unwrap() {
+                break;
+            }
+        }
+        assert_eq!(q.next_avail(), 7);
+        assert_eq!(q.next_used(), 7);
+    }
+
+    #[test]
+    fn test_invalid_avail_idx() {
+        // This is a negative test for the following MUST from the spec: `A driver MUST NOT
+        // decrement the available idx on a virtqueue (ie. there is no way to “unexpose” buffers).`.
+        // We validate that for this misconfiguration, the device does not panic.
+        let mem = &GuestMemoryMmap::<()>::from_ranges(&[(GuestAddress(0), 0x10000)]).unwrap();
+        let vq = MockSplitQueue::new(mem, 16);
+
+        let mut q: QueueState = vq.create_queue();
+
+        // q is currently valid.
+        assert!(q.is_valid(mem));
+
+        // The chains are (0, 1), (2, 3, 4), (5, 6).
+        for i in 0..7 {
+            let flags = match i {
+                1 | 4 | 6 => 0,
+                _ => VIRTQ_DESC_F_NEXT,
+            };
+
+            let desc = Descriptor::new((0x1000 * (i + 1)) as u64, 0x1000, flags, i + 1);
+            vq.desc_table().store(i, desc).unwrap();
+        }
+
+        vq.avail().ring().ref_at(0).unwrap().store(u16::to_le(0));
+        vq.avail().ring().ref_at(1).unwrap().store(u16::to_le(2));
+        vq.avail().ring().ref_at(2).unwrap().store(u16::to_le(5));
+        // Let the device know it can consume chains with the index < 2.
+        vq.avail().idx().store(u16::to_le(3));
+        // No descriptor chains are consumed at this point.
+        assert_eq!(q.next_avail(), 0);
+        assert_eq!(q.next_used(), 0);
+
+        loop {
+            q.disable_notification(mem).unwrap();
+
+            while let Some(chain) = q.iter(mem).unwrap().next() {
+                // Process the descriptor chain, and then add entries to the
+                // used ring.
+                let head_index = chain.head_index();
+                let mut desc_len = 0;
+                chain.for_each(|d| {
+                    if d.flags() & VIRTQ_DESC_F_WRITE == VIRTQ_DESC_F_WRITE {
+                        desc_len += d.len();
+                    }
+                });
+                q.add_used(mem, head_index, desc_len).unwrap();
+            }
+            if !q.enable_notification(mem).unwrap() {
+                break;
+            }
+        }
+        // The next chain that can be consumed should have index 3.
+        assert_eq!(q.next_avail(), 3);
+        assert_eq!(q.avail_idx(mem, Ordering::Acquire).unwrap(), Wrapping(3));
+        assert_eq!(q.next_used(), 3);
+        assert_eq!(q.used_idx(mem, Ordering::Acquire).unwrap(), Wrapping(3));
+        assert!(q.lock().ready());
+
+        // Decrement `idx` which should be forbidden. We don't enforce this thing, but we should
+        // test that we don't panic in case the driver decrements it.
+        vq.avail().idx().store(u16::to_le(1));
+
+        loop {
+            q.disable_notification(mem).unwrap();
+
+            while let Some(_chain) = q.iter(mem).unwrap().next() {
+                // In a real use case, we would do something with the chain here.
+            }
+
+            if !q.enable_notification(mem).unwrap() {
+                break;
+            }
+        }
+    }
+}
